@@ -12,25 +12,26 @@
     "prepublishOnly": "npm run build"
   },
   "files": [
-    "dist"
+    "dist",
+    "types"
   ],
-  "repository": "https://github.com/cwxyz007/color-palette.git",
-  "author": "cwxyz <jie844067636@gmail.com>",
+  "repository": "https://github.com/0x-jerry/img-color-palette.git",
+  "author": "Jerry Wang <x.jerry.wang@gmail.com>",
   "license": "MIT",
-  "devDependencies": {
-    "@preact/preset-vite": "^2.0.0",
-    "@types/node": "^14.14.35",
-    "@typescript-eslint/parser": "^4.18.0",
-    "eslint": "^7.22.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.3.1",
-    "preact": "^10.5.13",
-    "typescript": "^4.2.3",
-    "vite": "^2.1.2"
-  },
   "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
+    "registry": "https://registry.npmjs.org/",
+    "access": "public"
+  },
+  "dependencies": {
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.14",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@vitejs/plugin-react": "^1.1.4",
+    "typescript": "^4.5.5",
+    "vite": "^2.7.13"
   }
 }
