@@ -1,5 +1,5 @@
 import './style.css'
-import { render } from 'preact'
+import { render } from 'react-dom'
 import { app } from './App'
 
 render(app(), document.getElementById('app')!)
