@@ -1,5 +1,5 @@
 import { defineConfig } from 'vite'
-import preact from '@preact/preset-vite'
+import react from '@vitejs/plugin-react'
 import path from 'path'
 
 // https://vitejs.dev/config/
@@ -7,7 +7,7 @@ export default defineConfig(({ mode }) => {
   const isBuildLib = mode === 'lib'
 
   return ({
-    plugins: [preact()],
+    plugins: [react()],
     base: './',
     publicDir: isBuildLib ? 'none' : 'public',
     build: {
